\name{BloodExample}
\docType{data}
\alias{BloodExample}
\title{Example data from blood cells}
\description{
    Example bulk and single cell data from blood cells for deconvolution.
}
\format{
    A list containing properties:
    \tabular{rlll}{
        Name                \tab Type       \tab Description\cr
        bulk                \tab matrix     \tab bulk expression data\cr
        singleCellExpr      \tab matrix     \tab single-cell expression data\cr
        singleCellLabels    \tab vector     \tab cell type label for individual cells\cr
    }
}

\keyword{datasets}