\name{DeconBenchmark-package}
\alias{DeconBenchmark-package}
\alias{DeconBenchmark}
\docType{package}
\title{
\packageTitle{DeconBenchmark}
}
\description{
\packageDescription{DeconBenchmark}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DeconBenchmark}
\packageIndices{DeconBenchmark}
This package provides a common interface to run >45 deconvolution methods.
Visit \url{https://github.com/tinnlab/DeconBenchmark} for list of supported methods and tutorials.
}
\author{
\packageAuthor{DeconBenchmark}

Maintainer: \packageMaintainer{DeconBenchmark}
}

\keyword{ package }
\keyword{ DeconBenchmark }
