% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{.generateReference.markers}
\alias{.generateReference.markers}
\title{Calculate markers from single cell data}
\usage{
.generateReference.markers(singleCellExpr, singleCellLabels, log2Threshold = 1)
}
\description{
Calculate markers from single cell data. This method is not to be called by users.
}
