% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getMethodDockerRepos}
\alias{.getMethodDockerRepos}
\title{Get docker image name}
\usage{
.getMethodDockerRepos(method)
}
\arguments{
\item{method}{Method name}
}
\value{
docker image name for the method
}
\description{
Get docker image name. This method is not to be called by users.
}
