% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.writeArgs}
\alias{.writeArgs}
\title{Write arguments to a h5 file}
\usage{
.writeArgs(
  h5file,
  bulk,
  nCellTypes = NULL,
  markers = NULL,
  isMethylation = F,
  seed = 1,
  singleCellExpr = NULL,
  singleCellLabels = NULL,
  singleCellSubjects = NULL,
  cellTypeExpr = NULL,
  sigGenes = NULL,
  signature = NULL
)
}
\description{
Write deconvolution arguments to a h5 file. This method is not to be called by users.
}
