% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMethodsInputs}
\alias{getMethodsInputs}
\title{Get required inputs for a list of methods}
\usage{
getMethodsInputs(methods, containerEngine = "docker", verbose = T)
}
\arguments{
\item{methods}{An array of method names}

\item{containerEngine}{A string indicating the container engine, must be `docker` or `singularity`. The default value is `docker`.}

\item{verbose}{A logical value (TRUE or FALSE) indicating whether to print the output of running processes.}
}
\value{
a list with names/keys are methods and values of each key are required parameters of the coressponding method.
}
\description{
Get required inputs for a list of methods
}
\examples{

requiredInputs <- getMethodsInputs(c("DeconRNASeq", "scaden"))
print(requiredInputs$DeconRNASeq)
print(requiredInputs$scaden)

}
