% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSupportedMethods}
\alias{getSupportedMethods}
\title{Get supported deconvolution methods}
\usage{
getSupportedMethods()
}
\value{
an array of supported methods.
}
\description{
Get supported deconvolution methods
}
\examples{
supportedMethods <- getSupportedMethods()
}
