% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{runDeconvolution}
\alias{runDeconvolution}
\title{Run deconvolution}
\usage{
runDeconvolution(
  methods,
  bulk,
  nCellTypes = NULL,
  markers = NULL,
  isMethylation = F,
  singleCellExpr = NULL,
  singleCellLabels = NULL,
  singleCellSubjects = NULL,
  cellTypeExpr = NULL,
  sigGenes = NULL,
  signature = NULL,
  seed = 1,
  matlabLicenseFile = NULL,
  timeout = 60 * 60 * 12,
  dockerArgs = c("--cpus=8.0", "-m=32G", "--memory-reservation=4G"),
  verbose = T
)
}
\arguments{
\item{methods}{An array of method names. See getSupportedMethods() for the list of supported methods.}

\item{bulk}{A matrix of bulk data. Rows are genes and columns are bulk samples.}

\item{nCellTypes}{The number of cell types.}

\item{markers}{A list with name/key are cell types and values of each key are markers for the coressponding cell type.}

\item{isMethylation}{A logical value (TRUE or FALSE) indicating whether the input data is methylation data.}

\item{singleCellExpr}{A matrix of single cell data. Rows are genes and columns are single cell samples.}

\item{singleCellLabels}{A array of single cell labels. The order of the labels must be the same as the order of the single cell samples.}

\item{singleCellSubjects}{A array of single cell subjects. The order of the subjects must be the same as the order of the single cell samples.}

\item{cellTypeExpr}{A matrix of cell type data. Rows are genes and columns are cell types. This matrix normally contains the same number of genes as bulk expression matrix (not to be confused with signature matrix).}

\item{sigGenes}{A list of significant genes.}

\item{signature}{A matrix of signature data. Rows are genes and columns are cell types. This matrix normally contains only marker genes (not to be confused with cell type expresison matrix).}

\item{seed}{An integer value for the seed of the random number generator.}

\item{matlabLicenseFile}{A path to a matlab license file. This file is required for methods that use matlab. The hostid of this license is the same as the hostid of the host machine. The user of this license must be `root`. Visit Matlab license center (https://www.mathworks.com/licensecenter/licenses) to obtain a license file.}

\item{timeout}{A timeout in seconds for the docker container. The default value is 12 hours.}

\item{dockerArgs}{A list of extra arguments for the docker container. The default values allow docker container to use maximum 8 cpus, reserve 4GB of memory and allow it to use up to 32GB of memory.}

\item{verbose}{A logical value (TRUE or FALSE) indicating whether to print the output of running processes.}
}
\value{
A list with names/keys are methods and values of each key is the deconvolution result for the corresponding methods. The deconvolution result for each method is a list with two elements: `$P` and `$S`. The `$P` is a matrix cell type proportions (samples x cell type) and the `$S` is a matrix of cell type signatures (genes x cell type).
}
\description{
Run deconvolution for a list of methods.
See getSupportedMethods() for the list of supported methods.
Each method is run in a separate docker container.
Each method has its own required inputs.
See getMethodsInputs() for the input of each method.
}
\examples{
\dontrun{
library(DeconBenchmark)

allSupportMethods <- getSupportedMethods() # Get the list of supported methods
print(allSupportMethods)
methodsToRun <- c("ReFACTor", "scaden", "CIBERSORT") # Select methods to run (must be in the list of supported methods)
requiredInputs <- getMethodsInputs(methodsToRun) # Get the required inputs for each method
print(requiredInputs) # list(ReFACTor = c("bulk", "nCellTypes"), scaden = c("bulk", "singleCellExpr", "singleCellLabels"), CIBERSORT = c("bulk", "signature"))

data(BloodExample) # Load example data
print(names(BloodExample)) # c("bulk", "singleCellExpr", "singleCellLabels")
bulk <- BloodExample$bulk

# Run AdRoit only
deconvolutionResult <- runDeconvolution(methods = "ReFACTor", bulk = bulk, nCellTypes = 8)
proportion <- deconvolutionResult$ReFACTor$P
print(head(proportion))

# Run ARIC only
singleCellExpr <- BloodExample$singleCellExpr
singleCellLabels <- BloodExample$singleCellLabels

deconvolutionResult <- runDeconvolution(methods = "scaden", bulk = bulk, singleCellExpr = singleCellExpr, singleCellLabels = singleCellLabels)
proportion <- deconvolutionResult$scaden$P
print(head(proportion))

# Run CIBERSORT only
reference <- generateReference(singleCellExpr, singleCellLabels, type="signature") # Generate reference

deconvolutionResult <- runDeconvolution(methods = "CIBERSORT", bulk = bulk, signature=reference$signature)
proportion <- deconvolutionResult$CIBERSORT$P
print(head(proportion))

# Run all three methods
deconvolutionResults <- runDeconvolution(methodsToRun, bulk = bulk, singleCellExpr = singleCellExpr, singleCellLabels = singleCellLabels, signature=reference$signature)
proportions <- lapply(deconvolutionResults, function(res) res$P)

print(head(proportions$ReFACTor))
print(head(proportions$CIBERSORT))
print(head(proportions$scaden))
}
}
