% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{.generateReference.markers}
\alias{.generateReference.markers}
\title{Calculate markers from single cell data}
\usage{
.generateReference.markers(singleCellExpr, singleCellLabels, log2Threshold = 1)
}
\arguments{
\item{singleCellExpr}{A matrix of single cell expression data.}

\item{singleCellLabels}{A vector of cell type labels.}

\item{log2Threshold}{The log2 threshold for marker selection.}
}
\value{
A list with names/keys are cell type labels and values are the corresponding markers.
}
\description{
Calculate markers from single cell data. This method is not to be called by users.
}
