% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMethodsInputs}
\alias{getMethodsInputs}
\title{Get required inputs for a list of methods}
\usage{
getMethodsInputs(methods, verbose = T)
}
\arguments{
\item{methods}{An array of method names}
}
\value{
a list with names/keys are methods and values of each key are required parameters of the coressponding method.
}
\description{
Get required inputs for a list of methods
}
\examples{
\dontrun{
requiredInputs <- getRequiredInputs(c("AdRoit", "ARIC"))
print(requiredInputs$AdRoit)
print(requiredInputs$ARIC)
}
}
